Rename some modules to talk about PlutusTx
