+++ plutus-tx-plugin/plutus-tx-plugin.cabal
@@ -25,29 +25,29 @@ flag development
 
 library
     exposed-modules:
-        Language.Plutus.Lift
-        Language.Plutus.CoreToPLC.Plugin
-        Language.Plutus.CoreToPLC.Builtins
-        Language.Plutus.CoreToPLC.Compiler.Error
+        Language.PlutusTx.Lift
+        Language.PlutusTx.Plugin
+        Language.PlutusTx.Builtins
+        Language.PlutusTx.Compiler.Error
     hs-source-dirs: src
     other-modules:
-        Language.Plutus.CoreToPLC.PLCTypes
-        Language.Plutus.CoreToPLC.PIRTypes
-        Language.Plutus.CoreToPLC.Utils
-        Language.Plutus.Lift.LiftPir
-        Language.Plutus.Lift.THUtils
-        Language.Plutus.Lift.Instances
-        Language.Plutus.CoreToPLC.Compiler.Binders
-        Language.Plutus.CoreToPLC.Compiler.Builtins
-        Language.Plutus.CoreToPLC.Compiler.Laziness
-        Language.Plutus.CoreToPLC.Compiler.Expr
-        Language.Plutus.CoreToPLC.Compiler.Names
-        Language.Plutus.CoreToPLC.Compiler.Kind
-        Language.Plutus.CoreToPLC.Compiler.Primitives
-        Language.Plutus.CoreToPLC.Compiler.Type
-        Language.Plutus.CoreToPLC.Compiler.Types
-        Language.Plutus.CoreToPLC.Compiler.Utils
-        Language.Plutus.CoreToPLC.Compiler.ValueRestriction
+        Language.PlutusTx.PLCTypes
+        Language.PlutusTx.PIRTypes
+        Language.PlutusTx.Utils
+        Language.PlutusTx.Lift.LiftPir
+        Language.PlutusTx.Lift.THUtils
+        Language.PlutusTx.Lift.Instances
+        Language.PlutusTx.Compiler.Binders
+        Language.PlutusTx.Compiler.Builtins
+        Language.PlutusTx.Compiler.Laziness
+        Language.PlutusTx.Compiler.Expr
+        Language.PlutusTx.Compiler.Names
+        Language.PlutusTx.Compiler.Kind
+        Language.PlutusTx.Compiler.Primitives
+        Language.PlutusTx.Compiler.Type
+        Language.PlutusTx.Compiler.Types
+        Language.PlutusTx.Compiler.Utils
+        Language.PlutusTx.Compiler.ValueRestriction
     default-language: Haskell2010
     default-extensions: ExplicitForAll ScopedTypeVariables
                         DeriveGeneric StandaloneDeriving DeriveLift
