+++ plutus-tx-plugin/src/Language/PlutusTx/Compiler/Expr.hs
@@ -4,38 +4,38 @@
 {-# LANGUAGE ViewPatterns      #-}
 
 -- | Functions for compiling GHC Core expressions into Plutus Core terms.
-module Language.Plutus.CoreToPLC.Compiler.Expr (convExpr, convExprWithDefs, convDataConRef) where
+module Language.PlutusTx.Compiler.Expr (convExpr, convExprWithDefs, convDataConRef) where
 
-import           Language.Plutus.CoreToPLC.Compiler.Binders
-import           Language.Plutus.CoreToPLC.Compiler.Builtins
-import           Language.Plutus.CoreToPLC.Compiler.Error
-import           Language.Plutus.CoreToPLC.Compiler.Laziness
-import           Language.Plutus.CoreToPLC.Compiler.Names
-import           Language.Plutus.CoreToPLC.Compiler.Primitives
-import           Language.Plutus.CoreToPLC.Compiler.Type
-import           Language.Plutus.CoreToPLC.Compiler.Types
-import           Language.Plutus.CoreToPLC.Compiler.Utils
-import           Language.Plutus.CoreToPLC.PIRTypes
-import           Language.Plutus.CoreToPLC.Utils
+import           Language.PlutusTx.Compiler.Binders
+import           Language.PlutusTx.Compiler.Builtins
+import           Language.PlutusTx.Compiler.Error
+import           Language.PlutusTx.Compiler.Laziness
+import           Language.PlutusTx.Compiler.Names
+import           Language.PlutusTx.Compiler.Primitives
+import           Language.PlutusTx.Compiler.Type
+import           Language.PlutusTx.Compiler.Types
+import           Language.PlutusTx.Compiler.Utils
+import           Language.PlutusTx.PIRTypes
+import           Language.PlutusTx.Utils
 
-import qualified CoreUtils                                     as GHC
-import qualified GhcPlugins                                    as GHC
-import qualified MkId                                          as GHC
-import qualified PrelNames                                     as GHC
+import qualified CoreUtils                              as GHC
+import qualified GhcPlugins                             as GHC
+import qualified MkId                                   as GHC
+import qualified PrelNames                              as GHC
 
-import qualified Language.PlutusIR                             as PIR
-import qualified Language.PlutusIR.Compiler.Definitions        as PIR
+import qualified Language.PlutusIR                      as PIR
+import qualified Language.PlutusIR.Compiler.Definitions as PIR
 import           Language.PlutusIR.Compiler.Names
-import qualified Language.PlutusIR.MkPir                       as PIR
-import qualified Language.PlutusIR.Value                       as PIR
+import qualified Language.PlutusIR.MkPir                as PIR
+import qualified Language.PlutusIR.Value                as PIR
 
-import qualified Language.PlutusCore                           as PLC
+import qualified Language.PlutusCore                    as PLC
 
 import           Control.Monad.Reader
 
-import qualified Data.ByteString.Lazy                          as BSL
-import           Data.List                                     (elem, elemIndex)
-import qualified Data.List.NonEmpty                            as NE
+import qualified Data.ByteString.Lazy                   as BSL
+import           Data.List                              (elem, elemIndex)
+import qualified Data.List.NonEmpty                     as NE
 import           Data.Traversable
 
 {- Note [System FC and System FW]
