+++ plutus-tx-plugin/src/Language/PlutusTx/Compiler/Expr.hs-boot
@@ -1,9 +1,9 @@
 {-# LANGUAGE FlexibleContexts  #-}
 
-module Language.Plutus.CoreToPLC.Compiler.Expr (convExpr, convExprWithDefs, convDataConRef) where
+module Language.PlutusTx.Compiler.Expr (convExpr, convExprWithDefs, convDataConRef) where
 
-import           Language.Plutus.CoreToPLC.Compiler.Types
-import           Language.Plutus.CoreToPLC.PIRTypes
+import           Language.PlutusTx.Compiler.Types
+import           Language.PlutusTx.PIRTypes
 
 import qualified GhcPlugins                               as GHC
