+++ plutus-tx-plugin/src/Language/PlutusTx/Compiler/Laziness.hs
@@ -1,18 +1,18 @@
 {-# LANGUAGE FlexibleContexts  #-}
 {-# LANGUAGE OverloadedStrings #-}
 -- | Simulating laziness.
-module Language.Plutus.CoreToPLC.Compiler.Laziness where
+module Language.PlutusTx.Compiler.Laziness where
 
-import {-# SOURCE #-} Language.Plutus.CoreToPLC.Compiler.Expr
-import {-# SOURCE #-} Language.Plutus.CoreToPLC.Compiler.Type
-import           Language.Plutus.CoreToPLC.Compiler.Types
-import           Language.Plutus.CoreToPLC.PIRTypes
+import {-# SOURCE #-} Language.PlutusTx.Compiler.Expr
+import {-# SOURCE #-} Language.PlutusTx.Compiler.Type
+import           Language.PlutusTx.Compiler.Types
+import           Language.PlutusTx.PIRTypes
 
-import qualified Language.PlutusIR                        as PIR
+import qualified Language.PlutusIR                as PIR
 
 import           Language.PlutusCore.Quote
 
-import qualified GhcPlugins                               as GHC
+import qualified GhcPlugins                       as GHC
 
 {- Note [Object- vs meta-language combinators]
 Many of the things we define as *meta*-langugage combinators (i.e. operations on terms) could be defined
