+++ plutus-tx-plugin/src/Language/PlutusTx/Compiler/Names.hs
@@ -2,25 +2,25 @@
 {-# LANGUAGE FlexibleContexts #-}
 
 -- | Functions for compiling GHC names into Plutus Core names.
-module Language.Plutus.CoreToPLC.Compiler.Names where
+module Language.PlutusTx.Compiler.Names where
 
 
-import           Language.Plutus.CoreToPLC.Compiler.Kind
-import {-# SOURCE #-} Language.Plutus.CoreToPLC.Compiler.Type
-import           Language.Plutus.CoreToPLC.Compiler.Types
-import           Language.Plutus.CoreToPLC.PLCTypes
+import           Language.PlutusTx.Compiler.Kind
+import {-# SOURCE #-} Language.PlutusTx.Compiler.Type
+import           Language.PlutusTx.Compiler.Types
+import           Language.PlutusTx.PLCTypes
 
-import qualified GhcPlugins                               as GHC
+import qualified GhcPlugins                       as GHC
 
-import qualified Language.PlutusCore                      as PLC
-import qualified Language.PlutusCore.MkPlc                as PLC
+import qualified Language.PlutusCore              as PLC
+import qualified Language.PlutusCore.MkPlc        as PLC
 import           Language.PlutusCore.Quote
 
 import           Language.PlutusIR.Compiler.Names
 
 import           Data.List
-import qualified Data.List.NonEmpty                       as NE
-import qualified Data.Map                                 as Map
+import qualified Data.List.NonEmpty               as NE
+import qualified Data.Map                         as Map
 
 lookupName :: Scope -> GHC.Name -> Maybe PLCVar
 lookupName (Scope ns _) n = Map.lookup n ns
