+++ plutus-tx-plugin/src/Language/PlutusTx/Compiler/Type.hs
@@ -5,7 +5,7 @@
 
 -- | Functions for compiling GHC types into PlutusCore types, as well as compiling constructors,
 -- matchers, and pattern match alternatives.
-module Language.Plutus.CoreToPLC.Compiler.Type (
+module Language.PlutusTx.Compiler.Type (
     convType,
     convKind,
     getDataCons,
@@ -17,32 +17,32 @@ module Language.Plutus.CoreToPLC.Compiler.Type (
     NewtypeCoercion (..),
     splitNewtypeCoercion) where
 
-import           Language.Plutus.CoreToPLC.Compiler.Binders
-import           Language.Plutus.CoreToPLC.Compiler.Builtins
-import           Language.Plutus.CoreToPLC.Compiler.Error
-import {-# SOURCE #-} Language.Plutus.CoreToPLC.Compiler.Expr
-import           Language.Plutus.CoreToPLC.Compiler.Kind
-import           Language.Plutus.CoreToPLC.Compiler.Laziness
-import           Language.Plutus.CoreToPLC.Compiler.Names
-import           Language.Plutus.CoreToPLC.Compiler.Types
-import           Language.Plutus.CoreToPLC.Compiler.Utils
-import           Language.Plutus.CoreToPLC.PIRTypes
+import           Language.PlutusTx.Compiler.Binders
+import           Language.PlutusTx.Compiler.Builtins
+import           Language.PlutusTx.Compiler.Error
+import {-# SOURCE #-} Language.PlutusTx.Compiler.Expr
+import           Language.PlutusTx.Compiler.Kind
+import           Language.PlutusTx.Compiler.Laziness
+import           Language.PlutusTx.Compiler.Names
+import           Language.PlutusTx.Compiler.Types
+import           Language.PlutusTx.Compiler.Utils
+import           Language.PlutusTx.PIRTypes
 
-import qualified GhcPlugins                                  as GHC
-import qualified Pair                                        as GHC
-import qualified PrelNames                                   as GHC
-import qualified TysPrim                                     as GHC
+import qualified GhcPlugins                             as GHC
+import qualified Pair                                   as GHC
+import qualified PrelNames                              as GHC
+import qualified TysPrim                                as GHC
 
-import qualified Language.PlutusIR                           as PIR
-import qualified Language.PlutusIR.Compiler.Definitions      as PIR
+import qualified Language.PlutusIR                      as PIR
+import qualified Language.PlutusIR.Compiler.Definitions as PIR
 import           Language.PlutusIR.Compiler.Names
-import qualified Language.PlutusIR.MkPir                     as PIR
+import qualified Language.PlutusIR.MkPir                as PIR
 
 import           Control.Monad.Reader
 
-import           Data.List                                   (reverse, sortBy)
-import qualified Data.List.NonEmpty                          as NE
-import qualified Data.Set                                    as Set
+import           Data.List                              (reverse, sortBy)
+import qualified Data.List.NonEmpty                     as NE
+import qualified Data.Set                               as Set
 import           Data.Traversable
 
 -- Types
