+++ plutus-tx-plugin/src/Language/PlutusTx/Compiler/ValueRestriction.hs
@@ -4,15 +4,15 @@
 {-# LANGUAGE OverloadedStrings #-}
 
 -- | Functions for dealing with the Plutus Core value restriction.
-module Language.Plutus.CoreToPLC.Compiler.ValueRestriction where
+module Language.PlutusTx.Compiler.ValueRestriction where
 
-import           Language.Plutus.CoreToPLC.Compiler.Error
-import           Language.Plutus.CoreToPLC.Compiler.Laziness
-import           Language.Plutus.CoreToPLC.Compiler.Types
-import           Language.Plutus.CoreToPLC.PIRTypes
+import           Language.PlutusTx.Compiler.Error
+import           Language.PlutusTx.Compiler.Laziness
+import           Language.PlutusTx.Compiler.Types
+import           Language.PlutusTx.PIRTypes
 
-import qualified Language.PlutusIR                           as PIR
-import qualified Language.PlutusIR.Value                     as PIR
+import qualified Language.PlutusIR                   as PIR
+import qualified Language.PlutusIR.Value             as PIR
 
 import           Control.Monad.Reader
