+++ plutus-tx/plutus-tx.cabal
@@ -19,7 +19,7 @@ source-repository head
 
 library
     exposed-modules:
-        Language.Plutus.TH
+        Language.PlutusTx.TH
     hs-source-dirs: src
     default-language: Haskell2010
     default-extensions: ExplicitForAll ScopedTypeVariables
