+++ plutus-use-cases/src/Language/Plutus/Runtime/TH.hs
@@ -32,11 +32,11 @@ module Language.Plutus.Runtime.TH(
     eqTx
     ) where
 
-import           Language.Haskell.TH (Exp, Q)
-import qualified Language.Plutus.TH  as Builtins
+import           Language.Haskell.TH  (Exp, Q)
+import qualified Language.PlutusTx.TH as Builtins
 import           Wallet.UTXO.Runtime
 
-import           Prelude             (Bool (..), Eq (..), Int, Maybe (..))
+import           Prelude              (Bool (..), Eq (..), Int, Maybe (..))
 
 -- | Logical AND
 and :: Q Exp
